pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prometheus-eval"
version = "0.1.0"
description = "Comprehensive framework for rigorous evaluation of LLM prompt effectiveness"
readme = "README.md"
authors = [
  { name = "Research Team", email = "research@example.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "tiktoken>=0.5.2",
    "torch>=2.1.0",
    "transformers>=4.37.0",
    "sentence-transformers>=2.3.0",
    "datasets>=2.16.0",
    "nltk>=3.8.1",
    "sacrebleu>=2.4.0",
    "bert-score>=0.3.13",
    "rouge-score>=0.1.2",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "docker>=7.0.0",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "jsonschema>=4.20.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "loguru>=0.7.2",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
    "tenacity>=8.2.3",
    "aiohttp>=3.9.0",
    "asyncio-throttle>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-timeout>=2.2.0",
    "black>=23.12.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "ipython>=8.18.0",
    "jupyter>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*", "docs*", "docker-images*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath = [
    ".",
    "src",
]
